Updated openstack/openstack Project: openstack/glance-specs  b6c8778ac1b3a609f7f19d101c256198257a01af
Add alt text for badges

The RST image directive takes an alt parameter that's used to supply an
alt attribute for the HTML image element that's generated during RST
to HTML conversion.  The alt text is useful for accessibility purposes
and is also displayed if the image source is unavailable when the HTML
is generated.  Because of the latter point, we can't rely on the
accessibility features of the SVG image, we need to maintain some info
here in the README.rst file.

Change-Id: I0dd53bf1f42a81a697c3b8201d01c5f045f16f51