Updated openstack/openstack Project: openstack-infra/nodepool  a7f90f1da5ed6db5baaca2b1b709d386ae6137a9
Have an ending line-feed on the generated id_rsa.pub file

Previously there was no ending line feed on the generated id_rsa.pub
file. This could cause a problem as that file is copied to become the
authorized_keys file and made it difficult to add additional keys to
the file.

Change-Id: I830dfa3a0c0f17a0c504567ec64569cb918a0a1c