Updated openstack/openstack Project: openstack/python-glanceclient  1505a47ff38a7bca50f32cd37002f50fa5561909
Add alt text for badges

The RST image directive takes an alt parameter that's used to supply an
alt attribute for the HTML image element that's generated during RST
to HTML conversion.  The alt text is useful for accessibility purposes
and is also displayed if the image source is unavailable when the HTML
is generated.  Because of the latter point, we can't rely on the
accessibility features of the SVG image, we need to maintain some info
here in the README.rst file.

Change-Id: I1ca4a4d84cbb87c696b98d2d1d14f9ef792fcff6