Updated openstack/openstack Project: openstack/glance_store  68437adb7b14a0c23cc1d8d71efd8a9d9942faab
Add alt text for badges

The RST image directive takes an alt parameter that's used to supply an
alt attribute for the HTML image element that's generated during RST
to HTML conversion.  The alt text is useful for accessibility purposes
and is also displayed if the image source is unavailable when the HTML
is generated.  Because of the latter point, we can't rely on the
accessibility features of the SVG image, we need to maintain some info
here in the README.rst file.

Change-Id: Ibd1ab0189aed56da9b95be28615d070750dcf76d