Updated openstack/openstack Project: openstack-infra/tripleo-ci  c98a47d05f8ab0c0384360698dbc2034f025478f
Temporarily disable ceph in the updates job

It looks like updates are currently breaking on a Ceph-related issue,
and since the multinode updates job doesn't use ceph and is passing
this should get the job working again.  Once it's back voting we
can figure out how to fix Ceph.

Change-Id: I08924b0be702530afcd4cb3b25083a15213b0077
Related-Bug: 1641653