Updated openstack/openstack Project: openstack-infra/devstack-gate  eb895ca90282019493c7889f57e8c4143468cfa9
Add devstack_gate_vars.yaml to hold varibles

We want to force users that re-use d-g roles to provide their own
variables, instead of having per-role defaults.
We will set the variables used by our playbooks on devstack_gate_vars.yaml,
and feed it to the playbooks via either vars_files or -e on the CLI.
For this initial change, add the BASE variable, which will be looked up from
envvars and default to /opt/stack if not defined.

Change-Id: I9b6b76fbe9866d29327a0c087051e10a80ef55eb