Updated openstack/openstack Project: openstack/nova  32b5c8254ddd459e17ab5929fc7097a06cf7d2e0
libvirt: Test disk creation in test_hard_reboot

This test was asserting a call to _create_images_and_backing in a way
that didn't validate the data passed in any way. We update this to
instead validate that the expected disks are created.

This test highlights a bug in _create_images_and_backing, which is
passing a full path in place of a disk name. This will be fixed by
removing the call to _create_images_and_backing in a subsequent
commit.

This also allows us to validate the future removal of the call to
_create_images_and_backing.

Implements: bp/libvirt-imagebackend-refactor
Change-Id: I9d87c8c66f7a9ee15d0c26e8d514cd7b0eea1e09