Updated openstack/openstack Project: openstack/requirements  6db1dfdea979f429b1c1062f4543e0d3af455ecf
add dulwich for reno

Reno currently uses shell commands to access git history. Newer versions
of git change the output of git log in a way that eliminates information
reno needs to track file renames. Dulwich is a python library for
accessing git repo information. We plan to rewrite the scanner in reno
using dulwich to avoid issues with porcelain changes in the git command
line in the future.

We also reviewd gitpython, which uses the git CLI and is likely to have
the same issue, and pygit2, which uses libgit2 and is harder to pip
install. Dulwich has C extensions, but can be installed via pip without
external libraries from system packages.

Dulwich is released under the Apache 2.0 and GPLv2 licenses. We will
use the Apache license.

Dulwich supports Python 2.7 and 3.5.

Dulwich appears relatively actively maintained, with
updates from just under 2 weeks ago in the git repository
(https://github.com/jelmer/dulwich) and a release in October of this year.

Dulwich will not be part of the runtime dependencies of OpenStack
services, since it should only be used within reno. I believe that means
if we need to pin it, we can, but for now the global-requirements.txt
update does not include an upper bound.

Change-Id: Ib79d6ee6cc38752b3692ebcadbcbc34c9cae35a6
Related-Bug: #1633168
Signed-off-by: Doug Hellmann <doug@doughellmann.com>