Updated openstack/openstack Project: openstack/nova  4ed2575dad07af5852be76ace1148b03b3ae3f86
libvirt: Never copy a swap disk during cold migration

We previously didn't copy a swap disk if we were resizing it, and
instead recreated it on the remote end. With this change we never
copy a swap disk, even if we aren't resizing it. This is more
efficient and less surprising.

Implements: bp/libvirt-imagebackend-refactor
Change-Id: I2376c2f708e929426205e5ad59b12f8829dbd12b