Updated openstack/openstack Project: openstack/nova  5d1b4c8ce653b86db45b4f53ae3573e0c81d5a83
Fix pci_alias that include white spaces

When the PCI alias was getting translated to a PCI request,
all the white spaces were getting removed. So the pci_passthrough:alias
specified in the flavor never matched the device in pci_alias.
This patch fixes the bug by only stripping the white spaces
at the beginning or the end of the elements

Change-Id: I2888b03faf6c25bcb6011db40100d925972c2231
Closes-Bug: #1638200