Updated openstack/openstack Project: openstack/zaqar  127eaa2beca1c8495e6ce24db831546ed4b0c04d
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/6d0bfdfe1e2180c9bacb80a1f18cc066

Change-Id: Ib3ce121a6477c40faf3e190119f1b1985553cf5a