Updated openstack/openstack Project: openstack-infra/project-config  c5c65ca7369694714469a7f8ececd52e1ee37437
Move ovb-updates job from experimental to check non-voting

The ovb-updates job is now working and we want to get it running on
every patch.  Initially it is just added as non-voting to ensure it
is stable and to verify we have the capacity to run three ovb jobs
on rh1.

Change-Id: I56526e8d9bed481bbb2787b6d0770a6bf7208055