Updated openstack/openstack Project: openstack/governance  ae4f1c94a48a0268fe478d08a8ae854e40d7806d
Remove some projects from ironic governance

ironic-webclient no longer has contributors working on it, and will be
retired. It was never even close to production-ready.

python-dracclient is intended to only interact with a single vendor's
hardware, which the ironic team decided[0] we no longer wish to maintain
in ironic's governance. It will continue to be maintained outside of the
ironic governance.

python-wsmanclient is only used by our AMT driver, which is being
dropped from ironic governance soon.[1] Furthermore, it is code to speak
a particular flavor of XML, as I understand things, and so isn't really
part of ironic's mission. This will continue to be maintained outside of
ironic governance.

[0] https://review.openstack.org/#/c/392685/
[1] http://lists.openstack.org/pipermail/openstack-dev/2016-November/106862.html

Change-Id: I4fb9bfb69a75ed5aa0277959c88e6f2f753deebb