Updated openstack/openstack Project: openstack/governance  ba7665c99c0f5732a4bfc658d9659adac435b7b2
[badge] Add title elements to svg

Add title elements describing the svg for accessibility purposes.
This is of limited use because the svg will be brought into an html
document via the img element, but it could be useful if a visually
impaired user accesses the svg directly.  The title on each badge
is just a statement of the badge name.  While sparse, it's helpful
because while the text elements of each badge contain this info,
it's duplicated and split up and would not be easy to understand.

Change-Id: Ia0a9b09221cf6bce029d1a5cee523efef201cf54