Updated openstack/openstack Project: openstack-infra/tripleo-ci  dcc902cd8cfe81fdf6984c5410cb3a78fd9f31c9
Add line break between review link and duration

Because of the number of jobs we're reporting on the ci-status page
now, I see some weird line breaks in the middle of the duration.
This makes it difficult to visually scan job durations.  For example,
one job on the page currently looks like this:

397267,1 - 1h
52m log

Since there are already three lines in the first column of the
table, let's just make each job report two lines explicitly so they
don't get broken up in weird ways.  This change would result in the
following format for the job above:

397267,1
1h 52m log

Change-Id: Ia067ed16276a1b7ec4e08e636968467d5eacf752