Updated openstack/openstack Project: openstack/cinder  db82045fc0a32e0595e660c29151ef83d6584e71
Update attach_volume() with versionedobjects

We want to update both the attach_volume API and
detach_volume API to use volume versionedobjects.

There are three patches for it in the chain.

- Update attach_volume() with versionedobjects
- Update detach_volume() with versionedobjects
- Update test cases related to attach_volume()
  and detach_volume() in file
  cinder/tests/unit/test_volume.py

This patch updates attach_volume API to use
volume versionedobjects.  Changes were made to
be backwards compatible with older RPC clients.
It only includes changes to the core cinder code.
Changes in the drivers are left to each driver
maintainer to update.

Co-Authored-By: Thang Pham <thang.g.pham@gmail.com>
Co-Authored-By: lisali <xiaoyan.li@intel.com>
Partial-Implements: blueprint cinder-objects

Change-Id: I0ef693a9c3d72349563e7199aa27a0724f98a774