Updated openstack/openstack Project: openstack/cinder  6767593558a21a43b3d732787434c7d0d9cabeed
Update detach_volume() with versionedobjects

The following patch updates detach_volume
API to use volume versionedobjects.  Changes were
made to be backwards compatible with older RPC clients.
It only includes changes to the core cinder code.
Changes in the drivers are left to each driver
maintainer to update.

We want to update both the attach_volume()
and detach_volume() with versionedobjects.
There are 3 patches in the chain.

The patch sequence is:
- Update attach_volume() with versionedojbects
- Update detach_volume() with versionedobjects
- Update test cases related to attach_volume
  and detach_volume in file
  cinder/tests/unit/test_volume.py

Co-Authored-By: Thang Pham <thang.g.pham@gmail.com>
Co-Authored-By: lisali <xiaoyan.li@intel.com>
Partial-Implements: blueprint cinder-objects

Change-Id: Ib153342b20f5b3e74d2815cb422b411f5690443f