Updated openstack/openstack Project: openstack-infra/tripleo-ci  af9ad526454e48274d1cdf73d61936b0ab0aa719
Run log collection in parallel

Log collection takes a little while on each overcloud node, and
since most OVB nodes are going to be on separate compute nodes we
should be able to run those commands in parallel without any major
drawbacks (the one exception would be extract_logs which is run on
the undercloud, but that seems to be a very small part of the log
collection runtime).

This is done by starting the group of log collection commands in the
background and then using the 'wait' command to pause until they are
all finished.

Change-Id: If65e0e460467185a5695cf214272ef35e41c3f03