Updated openstack/openstack Project: openstack/ironic  ff3a8f1d7f432c7a1c44d422afe320376265ec8f
Support names for {cleaning,provisioning}_network

Names are easier to use for humans, so let's allow them.
The actual name -> UUID convertion happens during validation and its
result is cached for each interface instance.

This change will allow recreating a network without changing Ironic
configuration or setting a network name in advance (e.g. by an installer).

Note that when names are used, an administrator is responsible for ensuring
that no other networks have the same name (names are not unique in Neutron).

Change-Id: I9905bdc3f778310ba191b2bed6eb11bad11a2846
Closes-Bug: #1614938