Updated openstack/openstack Project: openstack-infra/project-config  948122bf123cd7948a3a05bf3f12a11ef1bd0634
Run tripleo-scenario(001,004) in Puppet OpenStack CI

Because Zuul v2.5 doesn't allow to mix:
- run a job if some files are touched
- run a job for a specific project

We need to duplicate the tripleo scenarios multinode jobs so:

- TripleO CI will continue to run classic jobs (when some files are
  touched in any tripleo project).
- Puppet OpenStack CI will run the same scenarios with a different name
  and different layout (project and branch specific).

This patch is a first iteration, starting with puppet-ceph to see how it
goes. If it works as expected, I'll send a patch for all other projects.

Change-Id: I83154fa99767ef150286bb7b336aea7cf866d6d3