Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  83114faba6b85bb0dae066ca3ec4f23871683fb5
Use standard python method naming for xml_config methods

Limit use of double underscore (or 'dunder') methods to python magic or
where it is an intention to avoid namespace collisions for classes
intended to be sub-classed.

An example were this would be applicable is a mixin class adding
memoisation to all methods would need to avoid namespace collisions
should it be added to a class intended to be sub-classed, and so for
any methods it uses to help manage the memoisation it should use
name-mangling to avoid them being accidentally overridden.

Change-Id: I0fe1e438687b23e2f0e1cef5229eef92e1a652ad