Updated openstack/openstack Project: openstack/horizon  93b1cef05019eefabd28e71e9af038274a14b9cf
Migrate Create Container to schema form

Remove custom form controller and modal and migrate
to the new schema form implementation.

Also, as a bonus, we now check the validity of the
new container name as the user enters it into the
form, before submission.

Add some basic docs with pointers and a simple
example.

Change-Id: I156ded96340c65710ee0aa9182f859243347e16b
Partially-Fixes: 1616306