Updated openstack/openstack Project: openstack/cinder  1e6f03dd823ef21bbdeda4671a1303c4743c8995
Correct the attributes of volume which created by clone a CG

If a volume in a consistency group is bootable or allow
multi-attach, after clone the consistency group, the created
volume is not bootable and allow multi-attach is set to False.
This patch will correct the attributes of volume according to
the source volume.

Change-Id: Icb4fad78cb218a134bb87e7d522829297cf93b94
Closes-Bug: #1498783