Updated openstack/openstack Project: openstack/trove  e85740a28b78cea149395f0d40181109059074f9
(fix troveclient gate) Use alt-demo network

This change uses the alt-demo network instead of sharing
the private network from the demo tenant.  This should
fix the tempest tests on python-troveclient complaining
about this and also brings us one step closer to using
neutron entirely.

Tested with the demo tenant as well and it still
puts on a network without using the --nic parameter, the
only drawback is that the ip doesn't appear on the trove
show command.

The scenario and api tests were modified to always pass
in the nic parameter (as it probably should have been done
from the beginning anyways).

Change-Id: I9f3cbae3490e9995ba5f835fc2304442b83464e4
Closes-Bug: 1647001