Updated openstack/openstack Project: openstack/heat  6f32733d514fd5a2f65128e0728e93cf93fc151b
Refactor, add encrypt/decrypt data dict functions to crypt

Just a refactor, no change in functionality.

The functions added to crypt are used to encrypt / decrypt resource
properties data dicts. Note that they should not be used for
encrypting / decrypting other things such as params or user creds
(which are just strings).  An intermediate json conversion of each
value in a dict takes place before it is encrypted/decrypted.

Change-Id: Id6bcc90cbf430095719315ac7e9d3e8c9e745012