Updated openstack/openstack Project: openstack/glance  6b72fc3ef242f8572a581a09d082fd29f5656dd0
Add alt text for badges

The RST image directive takes an alt parameter that's used to supply an
alt attribute for the HTML image element that's generated during RST
to HTML conversion.  The alt text is useful for accessibility purposes
and is also displayed if the image source is unavailable when the HTML
is generated.  Because of the latter point, we can't rely on the
accessibility features of the SVG image, we need to maintain some info
here in the README.rst file.

Change-Id: I598a99237b8581995e79be96172a42a752107357