Updated openstack/openstack Project: openstack/requirements  4f2ccf8c4cfde2152c490a1507c76051e718d337
Change the oslo vmware requirement

Networking-vSphere requires some changes in oslo vmware which is
available in oslo.vmware 2.15.0 release.

In order to change the Networking-vSphere requirement, we need to change
the openstack global-requirement also to use oslo.vmware 2.15.0 or later
version.

This patch set changes the global-requirement to use oslo.vmware 2.15.0
and later version.

Change-Id: I965e3809cd5f79e47dfa6465081c63dd616d6227