Updated openstack/openstack Project: openstack-dev/devstack  6d66e647ca35910cbca4d940b5203d7307efa1db
don't setup cells if n-cpu isn't also running

create_cell requires n-api and at least one n-cpu up and running. If
we have a configuration where it is not guarunteed that there is an
n-cpu at the end of a devstack run we have to skip this step and make
the user run it manually later.

Change-Id: I2287ab29f3c1a7252271dcce81673ef365615296