Updated openstack/openstack Project: openstack/tripleo-heat-templates  e46349a37c1e697aa1305349df8269586b5d517d
Use transient hostname for deployed servers

The name output returned by this template is expected to be the short
name rather than a FQDN.  Generally 'hostnamectl --static' returns a
FQDN and --transient will be the short name.

This change switches to using --transient and also simplifies the
script by dropping the unused outputs.

Change-Id: I19eaf9f66668f7e68765bad4018c0c60314f3f8f