Updated openstack/openstack Project: openstack/glance  036b1f4846799ed0c4741ee9a139c8363fb07329
Correct releasenote "Prepare for oslo.log 3.17.0"

The releasenote for change Ia054f8f637c826280722c0d2a6811fdbc0cf54ed
says a default value is changed in order to produce duplication of
logs, whereas the purpose of the change is to prevent duplication of
logs.  This patch corrects the release note.

Change-Id: I1bb8d6f234f483824b1e9eb67b96aeefb43a7885