Updated openstack/openstack Project: openstack/sahara  b9fecf9657ec018c6b551cd65382c6d3720aa13c
Adding tenant_id to regex_search

In order to allow the creation of clusters with the same across
projects, even when set to public, we need to filter only cluster from
the requester's project when creating a new cluster. This will narrow
the name check to same project only.

Change-Id: Idd43833c71ef486cf72ebe88ff9df495fff6b809
Partial-bug: #1584680