Updated openstack/openstack Project: openstack/cinder  eeeca836aa76428e92ad76042b329f606478a441
NetApp cDOT driver fails to clone from NFS cache

In some cases, the NetApp cDOT NFS driver fails to clone from an
NFS image cache. Root cause is code that compares extra specs of
the new volume to capability lists that the driver reports to the
scheduler (such as a list of disk types). It fails because a list
is mutable, so not hashable, so the code cannot do set-based
comparison operations against extra specs. The simple fix is to
stop doing set operations on extra specs. Also, update the time
on the image cache file so that our tests can distinguish between
a fast clone from cache and the slower copy-from-Glance fallback.

Change-Id: I1dfacd24d49b11967eeb0c6b8560114d01952d74
Closes-Bug: #1634203