Updated openstack/openstack Project: openstack-infra/project-config  69ddd326e098086b4c78e1afc7d8a70825f041b8
Run experimental os-log-analyze job on xenial

Part of Xenial transition. This depends on a devstack plugin, going
forward master branch of devstack will be tested on Xenial so we should
test this with Xenial as well. Did not do the trusty/xenial split with
this because os-log-analyze only has a master branch.

Finally moved the instantiation of the job template under os-loganalyze
instead of osprofiler (bad merge or mistake put it there maybe?).

Change-Id: Idc7221be734c28fe90bd5339e0d04d22d6d45775