Updated openstack/openstack Project: openstack/tempest  005972d364d000c542b60c5903280baa3d73b97a
Make common->compute->shelve_server param more clear

common->compute->shelve_server() function accept the
compute server client as first arg but it is not clearly named
or mentioned in doc string.

This will improve this function usage and will be helpful
if it will be defined as lib interfacein future.

Change-Id: I5b9a18b005f620d7eca2e52d8bbc90e9d4f95502