Updated openstack/openstack Project: openstack/cinder  85841a1713567e713160e10ea6004146d88aba63
VMware: VMDK driver performance improvements

Following optimizations are done to improve the performance of
volume create and attach operations:

* vCenter resource pool to datacenter references are cached to
avoid backend API calls to lookup datacenter for placing
volumes.

* Remove the expensive 'folder exists' check before creating a
folder in vCenter inventory to organize the volumes. This check
is unnecessary because we already have the logic to handle
duplicate folder exception.

* Skip datastore selection in _verify_volume_creation because
we repeat it before creating the volume in backend.

Also removing some of the redundant test cases.

Closes-bug: #1556902
Partial-bug: #1600754
Change-Id: I979c84b551cb045ae942a6427944bba9f17bfaa8