Updated openstack/openstack Project: openstack/nova  c49e05017b7fa96b7c3409ba33c7638ec5ba4c60
Fix nova-serialproxy when registering cli options

In latest devstack, nova-serialproxy fails to start because it crashed
when it tries to register the cli options.

The issue is that it tried to register an array of options by invoking
conf.register_cli_opt(), when multiple options need to be register
through conf.register_cli_opts().

Change-Id: Icb09f849290b78b5dd73c3c96a13ade02492ba3a
Closes-Bug: #1647784