Updated openstack/openstack Project: openstack/horizon  b71c1796dee4824939400b8f5257c09d9b156f20
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

b'https://gist.github.com/2b00c35799734733ecdec861c74975d8\n'

Change-Id: If27279ec325cefdf713dc924617429afc0834c87