Updated openstack/openstack Project: openstack/ironic-python-agent  9ebb70d2fdd75f888954ee8a9227ac3f38a90e67
Update mock variable name in unit tests

In test_agent.py change variable names from 'mocked_' to 'mock_' to
follow convention we use in openstack/ironic.

Change the mock variable names 'wsgi_server_cls' and
'mocked_server_maker' in some of the unit tests to
'mock_make_server'. It is the convention to have 'mock' in the
variable names for mock objects.

Not changing the other unit test files at this time.

Change-Id: I844071c928f92778e8ce0bfbdb8fb89fc26ee43b