Updated openstack/openstack Project: openstack/nova  852f40fd0c6e9d8878212ff3120556668023f1c4
Merge v21 view builder for ips API

Now v2 and v2.1 code is merged and we do not need
to maintain the different set of view builder.
Previously there were different way of building the
complete response for ips API. v2 used to do with extension
and v2.1 with view builder itself.

Part of bp:api-no-more-extensions

Change-Id: I438fdec60c221e2449203523e7c6fbae21b75604