Updated openstack/openstack Project: openstack/python-novaclient  814329655aee30a4c48f4a61d2720d5a5a8f9dc7
Revert "Microversion 2.39 - Simple tenant usage pagination"

This reverts commit f47393dbbfbf36616a8d28348d93e1368763b8c3.

The 2.39 microversion has not been merged in Nova, and this
change is missing a functional test to show the support added
here, which if present would have required a depends-on to the
Nova change for the 2.39 support. So this needs to come back
out until it's ready.

Change-Id: I6b6990b9a4e1e7cfea1d2aabe5f2464614440360