Updated openstack/openstack Project: openstack/heat  e4d7819a1df4282c38350a102fbeb5dbd2968c0a
Remove support for accessing ResourceDefinition like a dict

Heat no longer accesses resource definitions as if they were snippets of a
CloudFormation template, but rather uses the APIs of the ResourceDefinition
class. The ability to treat this class as a snippet of CloudFormation
template has been deprecated for several release cycles to allow
third-party plugins to move away from it; this change removes that ability.

Change-Id: Ia9c63662438d4c979b897118b3a7a736161bcbca