Updated openstack/openstack Project: openstack/heat  f67f2194d5ad9b78101ad78eb268a0f117a4cfe8
Remove the deprecated Stack.resolve_static_data() method

This method was left behind after the creation of the Function class and
resulting changes to function resolution in the Juno cycle, for the benefit
of any third-party plugins that may have been relying on it. It's time to
get rid of it.

Change-Id: Ia4700336f27ab259d25de9fd4a061c95198adf57