Updated openstack/openstack Project: openstack-infra/project-config  7b859a974ca663ea7907ae73f806638e7ab624b3
if a release has been tagged, stop processing it

When we were tagging releases by hand, sometimes the tag was applied
before any of the other processing was done. Now that all cycle-based
projects are using the release automation, we can skip updating
launchpad and the requirements repo if a release already exists.

Change-Id: I7c4a2c51dcba073b5b6d29ddf58fff1cb3f11055
Signed-off-by: Doug Hellmann <doug@doughellmann.com>