Updated openstack/openstack Project: openstack/glance-specs  ec48f0cb8171ff90e8399c86bd30a69e261e80fc
Use ``Range`` header in HTTP requests for partial downloads.

Currently, when a partial download request is sent, the
``Content-Range`` is parsed in glance/common/wsgi.py to
obtain the byte range. What this means is that the user
is expected to specify the byte range he/she intends to
download in a format of a HTTP range response.
For example: A user has to give "bytes 12-30/32" instead
of "bytes=12-32".
W3 standards (RFC 7233) specify ``Content-Range`` to be sent
as a response for a ``Range`` header in a HTTP request.
(https://tools.ietf.org/html/rfc7233#section-4.2)

This lite-spec suggests the acceptance of ``Range`` headers
for partial download requests.

Change-Id: I5bdadde682a0c50836bd95e2a6651d6e7e18f172