Updated openstack/openstack Project: openstack/python-ceilometerclient  bdd91689c84333ec1251e80d243da51d47048606
Don't include openstack/common in flake8 exclude list

There is no this directory openstack/common which was used
to keep codes from oslo-incubator, we have retired oslo-incubator.
Removing openstack/commonfrom all OpenStack code in favor of the
Oslo libraries is a project wide goal for the Ocata release. So
don't use this directory any more. We should drop it for improving
searching efficiency.

Change-Id: Ia154f7557575351173b9fe646e48bc5210ab698f