Updated openstack/openstack Project: openstack/heat  f141140bc85d23890e24698d084ad754247e9803
Use __slots__ for dependencies.Node

The number of nodes in the dependency graph is not huge (one per resource),
but we tend to make a lot of copies - the whole graph must be copied every
time we iterate through it. Since the Node class has only two attributes
and is basically never modified, we can easily save a little memory by
defining __slots__ for it.

Change-Id: Ibe36016a5889c9f1bd89672b1b03b2bc69b7925b