Updated openstack/openstack Project: openstack/nova  c106ff09d479c402549e8340b345358b45faa5bc
libvirt: Mock imagebackend template funcs in ImageBackendFixture

This represents a small change to how we test the arguments passed to
a template function. Most tests which test cache() currently don't
directly test the callback function. Some test the callback function
which was passed to cache(), but this is undesirable as:

  * It breaks untestably if you replace it with a wrapper
  * You can't test the arguments which were passed to it

To make this easier to test, and because a subsequent change alters
this slightly in ways we want to make obvious, we update
ImageBackendFixture to execute the callback function when cache() is
called. We pre-emptively mock all callback methods so they are not
actually called. Test can assert on these mocks to check that the
intended callback was called, and the arguments used.

Implements: bp/libvirt-imagebackend-refactor
Change-Id: Ifef488bcc3d8b7bb11aa7ec8255f9a27f57c3291