Updated openstack/openstack Project: openstack-infra/system-config  797c19af238bc0f6aba4bbc49d5d3109771140a8
Create new mailing list for Interop Working Group

On mandate from board, the DefCore Committee is changing its
name to the Interop Working Group. This change creates a new
mailing list with the updated name. Once the Interop Working
Group mailing list is created, the DefCore mailing list will
be shut down and left in place as an archive of previous
communication.

Change-Id: Iaf750d1887307c2686f20028368fd8eaa3534882